Refactor exception classes to the proper python package
