+++ sdmx2jsonld/sdmxattributes/confirmationStatus.py
@@ -20,7 +20,7 @@
 # under the License.
 ##
 from re import search
-from sdmx2jsonld.sdmxattributes.exceptions import ClassConfStatusError
+from sdmx2jsonld.exceptions.exceptions import ClassConfStatusError
 from sdmx2jsonld.common.commonclass import CommonClass
