+++ sdmx2jsonld/sdmxconcepts/refareaconcept.py
@@ -19,8 +19,6 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 ##
-from re import search
-from sdmx2jsonld.sdmxattributes.exceptions import ClassFreqError
 from sdmx2jsonld.common.commonclass import CommonClass
