+++ sdmx2jsonld/sdmxdimensions/timeperiod.py
@@ -20,7 +20,7 @@
 # under the License.
 ##
 from re import search
-from sdmx2jsonld.sdmxattributes.exceptions import ClassFreqError
+from sdmx2jsonld.exceptions.exceptions import ClassFreqError
 from sdmx2jsonld.common.commonclass import CommonClass
